rearrange tests a bit for convention and readability
